refactor: Move all config files to config folder
* Rename tests to test
* Keep project structure tidy
