+++ config/commitlint.config.js
