+++ config/qingcloud.commit.rules.js
