+++ src/i18n.js
@@ -1,7 +1,7 @@
 import i18n from 'i18next';
 // import Backend from 'i18next-xhr-backend';
 import LngDetector from 'i18next-browser-languagedetector';
-import baseConf from '../i18n.config';
+import baseConf from '../config/i18n.config';
 
 const lngDetectorOptions = {
   order: ['querystring', 'cookie', 'localStorage', 'navigator'],
