+++ test/components/Base/Switch.test.js
