+++ webpack.dev.js
@@ -1,6 +1,6 @@
 const { resolve } = require('path');
 const webpack = require('webpack');
-const postCssOptions = require('./postcss.options');
+const postCssOptions = require('./config/postcss.options');
 
 module.exports = {
   mode: 'development',
